Add files via upload
An example of JavaFX. Simply create a text while for this program to analyze that lists on the first line, the number of people you want analyzed. Then list their names on each line thereafter with their children listed to the right of each respective name, and this program will create a family tree representing the lineage.